Splice line node into staged function result tree
Also fixup the name of the lambda info after it is evaluated which is
normally done by jl_add_method. This should make it appear less wrong in
the profiler at least. It is arguably not the best possible thing since
this will point to the staged function itself and not where the
resulting expr was created. Still better than anonymous:???.